+++ spec/Pim/Bundle/CatalogBundle/Manager/ProductManagerSpec.php
@@ -132,10 +132,18 @@ class ProductManagerSpec extends ObjectBehavior
         $objectManager,
         ProductInterface $product
     ) {
-        $eventDispatcher->dispatch(
-            ProductEvents::PRE_REMOVE,
-            Argument::type('Symfony\Component\EventDispatcher\GenericEvent')
-        )->shouldBeCalled();
+        $eventDispatcher
+            ->dispatch(
+                ProductEvents::PRE_REMOVE,
+                Argument::type('Symfony\Component\EventDispatcher\GenericEvent')
+            )
+            ->shouldBeCalled();
+        $eventDispatcher
+            ->dispatch(
+                ProductEvents::POST_REMOVE,
+                Argument::type('Symfony\Component\EventDispatcher\GenericEvent')
+            )
+            ->shouldBeCalled();
 
         $objectManager->remove($product)->shouldBeCalled();
         $objectManager->flush()->shouldBeCalled();
@@ -148,10 +156,18 @@ class ProductManagerSpec extends ObjectBehavior
         $objectManager,
         ProductInterface $product
     ) {
-        $eventDispatcher->dispatch(
-            ProductEvents::PRE_REMOVE,
-            Argument::type('Symfony\Component\EventDispatcher\GenericEvent')
-        )->shouldBeCalled();
+        $eventDispatcher
+            ->dispatch(
+                ProductEvents::PRE_REMOVE,
+                Argument::type('Symfony\Component\EventDispatcher\GenericEvent')
+            )
+            ->shouldBeCalled();
+        $eventDispatcher
+            ->dispatch(
+                ProductEvents::POST_REMOVE,
+                Argument::type('Symfony\Component\EventDispatcher\GenericEvent')
+            )
+            ->shouldBeCalled();
 
         $objectManager->remove($product)->shouldBeCalled();
         $objectManager->flush()->shouldNotBeCalled();
@@ -170,6 +186,10 @@ class ProductManagerSpec extends ObjectBehavior
 
         $eventDispatcher->dispatch(
             ProductEvents::PRE_REMOVE,
+            Argument::type('Symfony\Component\EventDispatcher\GenericEvent')
+        )->shouldBeCalledTimes(2);
+        $eventDispatcher->dispatch(
+            ProductEvents::POST_REMOVE,
             Argument::type('Symfony\Component\EventDispatcher\GenericEvent')
         )->shouldBeCalledTimes(2);