+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/ProductRepository.php
@@ -3,7 +3,6 @@
 namespace Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM;
 
 use Doctrine\ODM\MongoDB\DocumentRepository;
-use Doctrine\ODM\MongoDB\Query\Expr;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ODM\MongoDB\Query\Builder as QueryBuilder;
 use Pim\Bundle\CatalogBundle\Repository\ReferableEntityRepositoryInterface;
@@ -683,5 +682,42 @@ class ProductRepository extends DocumentRepository implements
         }
 
         return $qb;
+    }
+
+    /**
+     * @param integer $productId
+     * @param integer $assocTypeCount
+     *
+     * @TODO: Make some refactoring with PublishedProductRepository
+     */
+    public function removeAssociatedProduct($productId, $assocTypeCount)
+    {
+        $mongoRef = [
+            '$ref' => $this->dm->getClassMetadata($this->documentName)->getCollection(),
+            '$id' => new \MongoId($productId),
+            '$db' => $this->dm->getConfiguration()->getDefaultDB(),
+        ];
+
+        $collection = $this->dm->getDocumentCollection($this->documentName);
+
+        // we iterate over the number of association types because the query removes only the product that
+        // belongs to the first association (instead of removing it in existing associations)
+        for ($i = 0; $i < $assocTypeCount; $i++) {
+            $collection->update(
+                [
+                    'associations' => [
+                        '$elemMatch' => [
+                            'products' => $mongoRef
+                        ]
+                    ]
+                ],
+                [
+                    '$pull' => [
+                        'associations.$.products' => $mongoRef
+                    ]
+                ],
+                [ 'multiple' => 1 ]
+            );
+        }
     }
 }