+++ src/Pim/Bundle/CatalogBundle/Event/ProductEvents.php
@@ -40,4 +40,34 @@ final class ProductEvents
      * @staticvar string
      */
     const PRE_REMOVE = 'pim_catalog.pre_remove.product';
+
+    /**
+     * This event is thrown after a product has been removed.
+     *
+     * The event listener receives an
+     * Symfony\Component\EventDispatcher\GenericEvent instance.
+     *
+     * @staticvar string
+     */
+    const POST_REMOVE = 'pim_catalog.post_remove.product';
+
+    /**
+     * This event is thrown before some products are removed
+     *
+     * The event listener receives an
+     * Symfony\Component\EventDispatcher\GenericEvent instance.
+     *
+     * @staticvar string
+     */
+    const PRE_MASS_REMOVE = 'pim_catalog.pre_mass_remove.product';
+
+    /**
+     * This event is thrown after some products have been removed.
+     *
+     * The event listener receives an
+     * Symfony\Component\EventDispatcher\GenericEvent instance.
+     *
+     * @staticvar string
+     */
+    const POST_MASS_REMOVE = 'pim_catalog.post_mass_remove.product';
 }