+++ src/Pim/Bundle/DataGridBundle/Extension/MassAction/Handler/DeleteProductsMassActionHandler.php
@@ -1,0 +1,81 @@
+<?php
+
+namespace Pim\Bundle\DataGridBundle\Extension\MassAction\Handler;
+
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Symfony\Component\EventDispatcher\GenericEvent;
+use Symfony\Component\Translation\TranslatorInterface;
+use Oro\Bundle\DataGridBundle\Datagrid\DatagridInterface;
+use Oro\Bundle\DataGridBundle\Extension\MassAction\Actions\MassActionInterface;
+use Oro\Bundle\DataGridBundle\Extension\MassAction\MassActionResponse;
+use Pim\Bundle\CatalogBundle\Event\ProductEvents;
+use Pim\Bundle\CatalogBundle\Manager\ProductManager;
+use Pim\Bundle\DataGridBundle\Datasource\ResultRecord\HydratorInterface;
+use Pim\Bundle\DataGridBundle\Extension\MassAction\Event\MassActionEvent;
+use Pim\Bundle\DataGridBundle\Extension\MassAction\Event\MassActionEvents;
+
+/**
+ * Mass delete products action handler
+ *
+ * @author    Romain Monceau <romain@akeneo.com>
+ * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class DeleteProductsMassActionHandler extends DeleteMassActionHandler
+{
+    /** @var ProductManager */
+    protected $productManager;
+
+    /**
+     * @param HydratorInterface        $hydrator
+     * @param TranslatorInterface      $translator
+     * @param EventDispatcherInterface $eventDispatcher
+     * @param ProductManager           $productManager
+     */
+    public function __construct(
+        HydratorInterface $hydrator,
+        TranslatorInterface $translator,
+        EventDispatcherInterface $eventDispatcher,
+        ProductManager $productManager
+    ) {
+        parent::__construct($hydrator, $translator, $eventDispatcher);
+
+        $this->productManager = $productManager;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * Dispatch two more events for products (PRE_MASS_REMOVE and POST_MASS_REMOVE)
+     */
+    public function handle(DatagridInterface $datagrid, MassActionInterface $massAction)
+    {
+        // dispatch pre handler event
+        $massActionEvent = new MassActionEvent($datagrid, $massAction, array());
+        $this->eventDispatcher->dispatch(MassActionEvents::MASS_DELETE_PRE_HANDLER, $massActionEvent);
+
+        $datasource = $datagrid->getDatasource();
+        $datasource->setHydrator($this->hydrator);
+
+        // hydrator uses index by id
+        $objectIds = $datasource->getResults();
+
+        try {
+            $this->eventDispatcher->dispatch(ProductEvents::PRE_MASS_REMOVE, new GenericEvent($objectIds));
+
+            $countRemoved = $datasource->getMassActionRepository()->deleteFromIds($objectIds);
+
+            $this->eventDispatcher->dispatch(ProductEvents::POST_MASS_REMOVE, new GenericEvent($objectIds));
+        } catch (\Exception $e) {
+            $errorMessage = $e->getMessage();
+
+            return new MassActionResponse(false, $this->translator->trans($errorMessage));
+        }
+
+        // dispatch post handler event
+        $massActionEvent = new MassActionEvent($datagrid, $massAction, $objectIds);
+        $this->eventDispatcher->dispatch(MassActionEvents::MASS_DELETE_POST_HANDLER, $massActionEvent);
+
+        return $this->getResponse($massAction, $countRemoved);
+    }
+}