+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed a bug that allowed to mass edit only 10 families
 - Fixed category order in the categories tab of products
 - Incomplete archives no longer appear as downloadable in the export execution details page
+- Fixed Cascade delete on associations for MongoDB impl
 
 ## BC breaks
 - Replace ACLs `pim_enrich_family_add_atribute` and `pim_enrich_family_remove_atribute` with `pim_enrich_family_edit_attributes`. This ACL also enforces rights to edit attribute requirements.